Fixed reading floats with single instead of double precision. Added missing geometry.cpp to git. Compare function will now import the entire file, not just record the differences once. Implemented colors. Fixed MDX padding. Changed 'Mass' to 'Batch'. Removed head link, game and platform options from the menu, making them buttons in the client area instead. Reimplemented reading array8 padding. TextureNumber is now calculated according to the MdxDataBitmap. Fixed import of inv_count.

+++ file.cpp
@@ -251,7 +251,7 @@ bool TextFile::ReadFloat(double & fNew, std::string * sGet, bool bPrint){
     if(sGet != nullptr) *sGet = sCheck;
 
     try{
-        fNew = std::stof(sCheck, (size_t*) NULL);
+        fNew = std::stod(sCheck, (size_t*) NULL);
     }
     catch(std::invalid_argument){
         std::cout << "TextFile::ReadFloat(): There was an error converting the string: " << sCheck << ". \n";
